MWPW-141920 refactor legacy-analytics (#934)
* MWPW-141920 refactor legacy-analytics

* MWPW-141920 refactor branch links

* MWPW-141920 if martech not ready wait for frictionless event

* MWPW-141920 make events fire safer

* MWPW-141920 set data analytic attributes for martech before loading script

* Use placeholders and Distinct BranchLinks from SearchBranchLinks (#936)

* use placeholders and distinct bl from sbl

* finalize search branch params

* stricter check

---------
