+++ express/blocks/faq/faq.js
@@ -2,6 +2,7 @@ import {
   createTag,
   getMetadata,
 } from '../../scripts/utils.js';
+import { trackButtonClick } from '../../scripts/instrument.js';
 
 function decorateFAQBlocks(block) {
   const showSchema = getMetadata('show-faq-schema');
@@ -22,6 +23,10 @@ function decorateFAQBlocks(block) {
     const { question, answer } = faq;
 
     const $accordion = createTag('div', { class: 'faq-accordion' });
+    // for tracking the faq
+    $accordion.addEventListener('click', () => {
+      trackButtonClick($accordion);
+    });
     block.append($accordion);
 
     const $questionDiv = createTag('h3', { class: 'faq-question' });
