+++ express/blocks/sticky-promo-bar/sticky-promo-bar.js
@@ -1,5 +1,6 @@
 import { createTag } from '../../scripts/utils.js';
 import BlockMediator from '../../scripts/block-mediator.min.js';
+import { sendEventToAnalytics } from '../../scripts/instrument.js';
 
 function initScrollInteraction(block) {
   const inBodyBanner = block.cloneNode(true);
@@ -44,6 +45,7 @@ export default async function decorate(block) {
       block,
       rendered: false,
     });
+    sendEventToAnalytics('adobe.com:express:cta:startYourFreeTrial:close');
   });
 
   if (block.classList.contains('loadinbody')) {
