+++ express/blocks/toggle-bar/toggle-bar.js
@@ -1,5 +1,6 @@
 import { createTag } from '../../scripts/utils.js';
 import { addTempWrapper } from '../../scripts/decorate.js';
+import { sendEventToAnalytics, textToName } from '../../scripts/instrument.js';
 
 function decorateButton(block, toggle) {
   const button = createTag('button', { class: 'toggle-bar-button' });
@@ -29,6 +30,21 @@ function decorateButton(block, toggle) {
 
   button.append(iconsWrapper, textWrapper);
   toggle.parentNode.replaceChild(button, toggle);
+
+  let texts = [];
+  let child = textWrapper.firstChild;
+  while (child) {
+    if (child.nodeType === 3) {
+      texts.push(child.data);
+    }
+    child = child.nextSibling;
+  }
+
+  texts = texts.join('') || textWrapper.textContent.trim();
+  const eventName = `adobe.com:express:homepage:intentToggle:${textToName(texts)}`;
+  button.addEventListener('click', () => {
+    sendEventToAnalytics(eventName);
+  });
 }
 
 function initButton(block, sections, index, props) {
