+++ express/scripts/express-delayed.js
@@ -109,9 +109,9 @@ export async function getProfile() {
   });
 }
 
-const branchLinkOrigins = ['https://adobesparkpost.app.link', 'https://adobesparkpost-web.app.link'];
+const branchLinkOriginPattern = /^https:\/\/adobesparkpost(-web)?\.app\.link/;
 function isBranchLink(url) {
-  return branchLinkOrigins.includes(new URL(url).origin);
+  return branchLinkOriginPattern.test(new URL(url).origin);
 }
 // product entry prompt
 async function canPEP() {
