+++ express/scripts/instrument.js
@@ -6,16 +6,78 @@ import {
   checkTesting,
   getAssetDetails,
   getMetadata,
-  fetchPlaceholders,
 } from './utils.js';
+import trackBranchParameters from './branchlinks.js';
 
 const usp = new URLSearchParams(window.location.search);
 const martech = usp.get('martech');
 const w = window;
+if (!w.alloy_all) w.alloy_all = {};
 const d = document;
 const loc = w.location;
 const { pathname } = loc;
 let sparkLandingPageType;
+let sparkTouchpoint;
+// home
+if (
+  pathname === '/express'
+  || pathname === '/express/'
+) {
+  sparkLandingPageType = 'home';
+  // seo
+} else if (
+  pathname === '/express/create'
+  || pathname.includes('/create/')
+  || pathname === '/express/make'
+  || pathname.includes('/make/')
+  || pathname === '/express/feature'
+  || pathname.includes('/feature/')
+  || pathname === '/express/discover'
+  || pathname.includes('/discover/')
+) {
+  sparkLandingPageType = 'seo';
+  // learn
+} else if (
+  pathname === '/express/tools'
+  || pathname.includes('/tools/')
+) {
+  sparkLandingPageType = 'quickAction';
+} else if (
+  pathname === '/express/learn'
+  || (
+    pathname.includes('/learn/')
+    && !pathname.includes('/blog/')
+  )
+) {
+  if (pathname.includes('/express-your-brand')) {
+    sparkLandingPageType = 'express-your-brand';
+  } else {
+    sparkLandingPageType = 'learn';
+  }
+  // blog
+} else if (
+  pathname === '/express/learn/blog'
+  || pathname.includes('/learn/blog/')
+) {
+  sparkLandingPageType = 'blog';
+  // pricing
+} else if (
+  pathname.includes('/pricing')
+) {
+  sparkLandingPageType = 'pricing';
+  // edu
+} else if (
+  pathname.includes('/education/')
+) {
+  sparkLandingPageType = 'edu';
+  // other
+} else if (
+  pathname.includes('/express-your-fandom')
+) {
+  sparkLandingPageType = 'express-your-fandom';
+} else {
+  sparkLandingPageType = 'other';
+}
 
 // alloy feature flag
 let martechURL;
@@ -57,30 +119,8 @@ window.marketingtech = {
 // w.targetGlobalSettings = w.targetGlobalSettings || {};
 // w.targetGlobalSettings.bodyHidingEnabled = checkTesting();
 
-function getPlacement(btn) {
-  const parentBlock = btn.closest('.block');
-  let placement = 'outside-blocks';
-
-  if (parentBlock) {
-    const blockName = parentBlock.dataset.blockName || parentBlock.classList[0];
-    const sameBlocks = btn.closest('main')?.querySelectorAll(`.${blockName}`);
-
-    if (sameBlocks && sameBlocks.length > 1) {
-      sameBlocks.forEach((b, i) => {
-        if (b === parentBlock) {
-          placement = `${blockName}-${i + 1}`;
-        }
-      });
-    } else {
-      placement = blockName;
-    }
-
-    if (['template-list', 'template-x'].includes(blockName) && btn.classList.contains('placeholder')) {
-      placement = 'blank-template-cta';
-    }
-  }
-
-  return placement;
+export function getExpressLandingPageType() {
+  return sparkLandingPageType;
 }
 
 function set(path, value) {
@@ -102,194 +142,133 @@ function set(path, value) {
   return obj;
 }
 
-export function sendEventToAdobeAnaltics(eventName) {
-  _satellite.track('event', {
-    xdm: {},
-    data: {
-      eventType: 'web.webinteraction.linkClicks',
-      web: {
-        webInteraction: {
-          name: eventName,
-          linkClicks: {
-            value: 1,
-          },
-          type: 'other',
-        },
-      },
-      _adobe_corpnew: {
-        digitalData: {
-          primaryEvent: {
-            eventInfo: {
-              eventName,
-            },
-          },
-        },
-      },
-    },
-  });
+function setDataAnalyticsAttributesForMartech() {
+  /* eslint-disable no-underscore-dangle */
+  //------------------------------------------------------------------------------------
+  // gathering the data
+  //------------------------------------------------------------------------------------
+
+  const locale = getConfig().locale.prefix;
+  const pathSegments = pathname.substr(1).split('/');
+  if (locale !== '') pathSegments.shift();
+  const pageName = `adobe.com:${pathSegments.join(':')}`;
+
+  const language = document.documentElement.getAttribute('lang');
+
+  let category = getMetadata('category');
+  if (!category && (pathname.includes('/create/')
+    || pathname.includes('/feature/'))) {
+    category = 'design';
+    if (pathname.includes('/image')) category = 'photo';
+    if (pathname.includes('/video')) category = 'video';
+  }
+
+  const url = new URL(loc.href);
+  sparkTouchpoint = url.searchParams.get('touchpointName');
+
+  //----------------------------------------------------------------------------
+  // set some global and persistent data layer properties
+  //----------------------------------------------------------------------------
+  set('page.pageInfo.pageName', pageName);
+  set('page.pageInfo.language', language);
+  set('page.pageInfo.siteSection', 'adobe.com:express');
+  set('page.pageInfo.category', category);
+
+  //----------------------------------------------------------------------------
+  // spark specific global and persistent data layer properties
+  //----------------------------------------------------------------------------
+  set('page.pageInfo.pageurl', loc.href);
+  set('page.pageInfo.namespace', 'express');
+
+  /* set experiment and variant information */
+  if (window.hlx.experiment) {
+    const { experiment } = window.hlx;
+    set('adobe.experienceCloud.target.info.primarytest.testinfo.campaignid', experiment.id);
+    set('adobe.experienceCloud.target.info.primarytest.testinfo.offerid', experiment.selectedVariant);
+  }
 }
 
-function sendFrictionlessEventToAdobeAnaltics(block) {
-  const eventName = 'view-quickaction-upload-page';
-  _satellite.track('event', {
-    xdm: {},
-    data: {
-      eventType: 'web.webinteraction.linkClicks',
-      web: {
-        webInteraction: {
-          name: eventName,
-          linkClicks: {
-            value: 1,
+function safelyFireAnalyticsEvent(event) {
+  // eslint-disable-next-line no-underscore-dangle
+  if (window._satellite?.track) {
+    event();
+  } else {
+    window.addEventListener('alloy_sendEvent', () => {
+      event();
+    }, { once: true });
+  }
+}
+
+export function sendEventToAnalytics(eventName) {
+  const fireEvent = () => {
+    _satellite.track('event', {
+      xdm: {},
+      data: {
+        eventType: 'web.webinteraction.linkClicks',
+        web: {
+          webInteraction: {
+            name: eventName,
+            linkClicks: {
+              value: 1,
+            },
+            type: 'other',
           },
-          type: 'other',
         },
-      },
-      _adobe_corpnew: {
-        sdm: {
-          event: {
-            pagename: eventName,
-            url: loc.href,
-          },
-          custom: {
-            qa: {
-              group: block.dataset.frictionlessgroup ?? 'unknown',
-              type: block.dataset.frictionlesstype ?? 'unknown',
+        _adobe_corpnew: {
+          digitalData: {
+            primaryEvent: {
+              eventInfo: {
+                eventName,
+              },
             },
           },
         },
       },
-    },
-  });
+    });
+  };
+  safelyFireAnalyticsEvent(fireEvent);
+}
+
+export function sendFrictionlessEventToAdobeAnaltics(block) {
+  const eventName = 'view-quickaction-upload-page';
+  const fireEvent = () => {
+    _satellite.track('event', {
+      xdm: {},
+      data: {
+        eventType: 'web.webinteraction.linkClicks',
+        web: {
+          webInteraction: {
+            name: eventName,
+            linkClicks: {
+              value: 1,
+            },
+            type: 'other',
+          },
+        },
+        _adobe_corpnew: {
+          sdm: {
+            event: {
+              pagename: eventName,
+              url: loc.href,
+            },
+            custom: {
+              qa: {
+                group: block.dataset.frictionlessgroup ?? 'unknown',
+                type: block.dataset.frictionlesstype ?? 'unknown',
+              },
+            },
+          },
+        },
+      },
+    });
+  };
+  safelyFireAnalyticsEvent(fireEvent);
 }
 
 export function textToName(text) {
   const splits = text.toLowerCase().split(' ');
   const camelCase = splits.map((s, i) => (i ? s.charAt(0).toUpperCase() + s.substr(1) : s)).join('');
   return (camelCase);
-}
-
-export async function trackBranchParameters($links) {
-  const placeholders = await fetchPlaceholders();
-  const rootUrl = new URL(window.location.href);
-  const params = rootUrl.searchParams;
-  const pageUrl = window.location.pathname;
-
-  const { experiment } = window.hlx;
-  const { referrer } = window.document;
-  const experimentStatus = experiment ? experiment.status.toLocaleLowerCase() : null;
-
-  const [
-    searchTerm,
-    canvasHeight,
-    canvasWidth,
-    canvasUnit,
-    sceneline,
-    taskID,
-    assetCollection,
-    category,
-    searchCategory,
-    loadPrintAddon,
-    tab,
-    action,
-    prompt,
-    sdid,
-    mv,
-    mv2,
-    sKwcId,
-    efId,
-    promoId,
-    trackingId,
-    cgen,
-  ] = [
-    getMetadata('branch-search-term'),
-    getMetadata('branch-canvas-height'),
-    getMetadata('branch-canvas-width'),
-    getMetadata('branch-canvas-unit'),
-    getMetadata('branch-sceneline'),
-    getMetadata('branch-task-id'),
-    getMetadata('branch-asset-collection'),
-    getMetadata('branch-category'),
-    getMetadata('branch-search-category'),
-    getMetadata('branch-loadprintaddon'),
-    getMetadata('branch-tab'),
-    getMetadata('branch-action'),
-    getMetadata('branch-prompt'),
-    params.get('sdid'),
-    params.get('mv'),
-    params.get('mv2'),
-    params.get('s_kwcid'),
-    params.get('ef_id'),
-    params.get('promoid'),
-    params.get('trackingid'),
-    params.get('cgen'),
-  ];
-
-  $links.forEach((a) => {
-    if (a.href && a.href.match('adobesparkpost.app.link')) {
-      a.rel = 'nofollow';
-      const btnUrl = new URL(a.href);
-      const isSearchBranchLink = placeholders['search-branch-links']?.replace(/\s/g, '').split(',').includes(`${btnUrl.origin}${btnUrl.pathname}`);
-      const urlParams = btnUrl.searchParams;
-      const setParams = (k, v) => {
-        if (v) urlParams.set(k, encodeURIComponent(v));
-      };
-      if (urlParams.has('acomx-dno')) {
-        urlParams.delete('acomx-dno');
-        btnUrl.search = urlParams.toString();
-        a.href = decodeURIComponent(btnUrl.toString());
-        return;
-      }
-      const placement = getPlacement(a);
-
-      if (isSearchBranchLink) {
-        setParams('category', category || 'templates');
-        setParams('taskID', taskID);
-        setParams('assetCollection', assetCollection);
-
-        if (searchCategory) {
-          setParams('searchCategory', searchCategory);
-        } else if (searchTerm) {
-          setParams('q', searchTerm);
-        }
-        if (loadPrintAddon) setParams('loadPrintAddon', loadPrintAddon);
-        setParams('tab', tab);
-        setParams('action', action);
-        setParams('prompt', prompt);
-      }
-
-      setParams('referrer', referrer);
-      setParams('url', pageUrl);
-      setParams('height', canvasHeight);
-      setParams('width', canvasWidth);
-      setParams('unit', canvasUnit);
-      setParams('sceneline', sceneline);
-      setParams('sdid', sdid);
-      setParams('mv', mv);
-      setParams('mv2', mv2);
-      setParams('efid', efId);
-      setParams('promoid', promoId);
-      setParams('trackingid', trackingId);
-      setParams('cgen', cgen);
-      setParams('placement', placement);
-
-      if (sKwcId) {
-        const sKwcIdParameters = sKwcId.split('!');
-
-        if (typeof sKwcIdParameters[2] !== 'undefined' && sKwcIdParameters[2] === '3') {
-          setParams('customer_placement', 'Google%20AdWords');
-        }
-
-        if (typeof sKwcIdParameters[8] !== 'undefined' && sKwcIdParameters[8] !== '') {
-          setParams('keyword', sKwcIdParameters[8]);
-        }
-      }
-
-      experimentStatus === 'active' && setParams('expid', `${experiment.id}-${experiment.selectedVariant}`);
-
-      btnUrl.search = urlParams.toString();
-      a.href = decodeURIComponent(btnUrl.toString());
-    }
-  });
 }
 
 export function appendLinkText(eventName, a) {
@@ -321,191 +300,194 @@ export function appendLinkText(eventName, a) {
 }
 
 export function trackButtonClick(a) {
-  let adobeEventName = 'adobe.com:express:cta:';
-  let hemingwayAssetId;
-  let hemingwayAssetPath;
-  let hemingwayAssetPosition;
+  const fireEvent = () => {
+    let adobeEventName = 'adobe.com:express:cta:';
+    let hemingwayAssetId;
+    let hemingwayAssetPath;
+    let hemingwayAssetPosition;
 
-  const hemingwayAsset = a.querySelector('picture,video,audio,img')
+    const hemingwayAsset = a.querySelector('picture,video,audio,img')
       || a.closest('[class*="-container"],[class*="-wrapper"]')?.querySelector('picture,video,audio,img');
-  const block = a.closest('.block');
-  const urlConstructable = a.href || a.currentSrc || a.src;
-  if (hemingwayAsset && block && urlConstructable) {
-    const { assetId, assetPath } = getAssetDetails(hemingwayAsset);
-    hemingwayAssetPath = assetPath;
-    hemingwayAssetId = assetId;
+    const block = a.closest('.block');
+    const urlConstructable = a.href || a.currentSrc || a.src;
+    if (hemingwayAsset && block && urlConstructable) {
+      const { assetId, assetPath } = getAssetDetails(hemingwayAsset);
+      hemingwayAssetPath = assetPath;
+      hemingwayAssetId = assetId;
 
-    const siblings = [...block
-      .querySelectorAll(`.${a.className.split(' ').join('.')}`)];
-    hemingwayAssetPosition = siblings.indexOf(a);
-  }
+      const siblings = [...block
+        .querySelectorAll(`.${a.className.split(' ').join('.')}`)];
+      hemingwayAssetPosition = siblings.indexOf(a);
+    }
 
-  const $templateContainer = a.closest('.template-list');
-  const $tutorialContainer = a.closest('.tutorial-card');
-  const $contentToggleContainer = a.closest('.content-toggle');
-  const $chooseYourPathContainer = a.closest('.choose-your-path');
-  // let cardPosition;
-  // Template button click
-  if ($templateContainer) {
-    adobeEventName += 'template:';
+    const $templateContainer = a.closest('.template-list');
+    const $tutorialContainer = a.closest('.tutorial-card');
+    const $contentToggleContainer = a.closest('.content-toggle');
+    const $chooseYourPathContainer = a.closest('.choose-your-path');
+    // let cardPosition;
+    // Template button click
+    if ($templateContainer) {
+      adobeEventName += 'template:';
 
-    const $img = a.querySelector('img');
+      const $img = a.querySelector('img');
 
-    // try to get the image alternate text
-    if (a.classList.contains('template-title-link')) {
-      adobeEventName += 'viewAll';
-    } else if (a.classList.contains('placeholder')) {
-      adobeEventName += 'createFromScratch';
-    } else if ($img && $img.alt) {
-      adobeEventName += textToName($img.alt);
+      // try to get the image alternate text
+      if (a.classList.contains('template-title-link')) {
+        adobeEventName += 'viewAll';
+      } else if (a.classList.contains('placeholder')) {
+        adobeEventName += 'createFromScratch';
+      } else if ($img && $img.alt) {
+        adobeEventName += textToName($img.alt);
+      } else {
+        adobeEventName += 'Click';
+      }
+      // Button in the FAQ
+    } else if ($tutorialContainer) {
+      const videoName = textToName(a.querySelector('h3').textContent.trim());
+      adobeEventName = `${adobeEventName}tutorials:${videoName}:tutorialPressed`;
+    } else if ($chooseYourPathContainer) {
+      const $slideTitle = a.querySelector('.choose-your-path-slide-title');
+      const slideName = $slideTitle ? textToName($slideTitle.textContent.trim()) : 'slide';
+
+      adobeEventName = `${adobeEventName}chooseYourPath:${slideName}:slidePressed`;
+    } else if ($contentToggleContainer) {
+      const toggleName = textToName(a.textContent.trim());
+      adobeEventName = `${adobeEventName}contentToggle:${toggleName}:buttonPressed`;
+    } else if (a.classList.contains('floating-button-lottie')) {
+      adobeEventName = `${adobeEventName}floatingButton:scrollPressed`;
+    } else if (a.classList.contains('video-player-inline-player-overlay')) {
+      const sessionName = a.parentNode.parentNode.parentNode.querySelector('.video-player-session-number').textContent.trim();
+      const videoName = a.parentNode.parentNode.parentNode.querySelector('.video-player-video-title').textContent.trim();
+      adobeEventName = `${adobeEventName}playing:${sessionName}-${videoName}`;
+    } else if (a.classList.contains('notch')) {
+      adobeEventName = `${adobeEventName}splitAction:notch`;
+    } else if (a.classList.contains('underlay')) {
+      adobeEventName = `${adobeEventName}splitAction:background`;
+    } else if (a.parentElement.classList.contains('floating-button')) {
+      adobeEventName = `${adobeEventName}floatingButton:ctaPressed`;
+    } else if (a.closest('.faq')) {
+      adobeEventName = appendLinkText(`${adobeEventName}faq:`, a);
+      // CTA in the hero
+    } else if (a.closest('.hero')) {
+      adobeEventName = appendLinkText(`${adobeEventName}hero:`, a);
+      // Click in the pricing block
+    } else if (sparkLandingPageType === 'express-your-fandom') {
+      adobeEventName = appendLinkText(`${adobeEventName}${sparkLandingPageType}:`, a);
+    } else if (sparkLandingPageType === 'express-your-brand') {
+      adobeEventName = appendLinkText(`${adobeEventName}learn:${sparkLandingPageType}:`, a);
+    } else if (sparkLandingPageType === 'pricing') {
+      if (a.tagName !== 'A') {
+        adobeEventName += `pricing:pricing:${a.textContent.trim()}:Click`;
+        // Creative cloud learn more
+      } else if (a.parentElement.id === 'adobe-spark-is-a-part-of-most-creative-cloud-paid-plans-learn-more') {
+        adobeEventName += 'pricing:creativeCloud:learnMore';
+      } else if (a.id === 'free-trial') {
+        adobeEventName += 'pricing:cta:StartForFree';
+      } else if (a.id === '3-month-trial') {
+        adobeEventName += 'pricing:cta:StartYour3MonthTrial';
+        // View plans
+      } else {
+        adobeEventName = 'adobe.com:express:CTA:pricing:viewPlans:Click';
+      }
+      // quick actions clicks
+    } else if (a.closest('ccl-quick-action') && a.classList.contains('upload-your-photo')) {
+      // this event is handled at mock-file-input level
+      return;
+    } else if (a.href && (a.href.match(/spark\.adobe\.com\/[a-zA-Z-]*\/?tools/g) || a.href.match(/express\.adobe\.com\/[a-zA-Z-]*\/?tools/g))) {
+      adobeEventName = appendLinkText(adobeEventName, a);
+    } else if (a.href && (a.href.match(/spark\.adobe\.com\/[a-zA-Z-]*\/?tools/g) || a.href.match(/express\.adobe\.com\/[a-zA-Z-]*\/?express-apps\/animate-from-audio/g))) {
+      adobeEventName = appendLinkText(adobeEventName, a);
+      // Frictionless Quick Actions clicks
+    } else if (a.closest('ccl-quick-action') && (a.getAttribute('data-action') === 'Download')) {
+      adobeEventName = 'quickAction:downloadPressed';
+    } else if (a.closest('ccl-quick-action') && (a.getAttribute('data-action') === 'Editor')) {
+      adobeEventName = 'quickAction:openInEditorPressed';
+      // ToC clicks
+    } else if (a.closest('.toc-container')) {
+      if (a.classList.contains('toc-toggle')) {
+        adobeEventName += 'toc:toggle:Click';
+      } else if (a.classList.contains('toc-close')) {
+        adobeEventName += 'toc:close:Click';
+      } else if (a.classList.contains('toc-handle')) {
+        adobeEventName += 'toc:close:Click:handle';
+      } else if (a.classList.contains('toc-wrapper')) {
+        adobeEventName += 'toc:close:Click:background';
+      } else {
+        adobeEventName = appendLinkText(`${adobeEventName}toc:link:Click:`, a);
+      }
+    } else if (a.closest('.template')) {
+      adobeEventName = appendLinkText(adobeEventName, a);
+    } else if (a.closest('.tabs-ax .tab-list-container')) {
+      adobeEventName += `${a.closest('.tabs-ax')?.id}:${a.id}`;
+      // Default clicks
     } else {
-      adobeEventName += 'Click';
+      adobeEventName = appendLinkText(adobeEventName, a);
     }
-    // Button in the FAQ
-  } else if ($tutorialContainer) {
-    const videoName = textToName(a.querySelector('h3').textContent.trim());
-    adobeEventName = `${adobeEventName}tutorials:${videoName}:tutorialPressed`;
-  } else if ($chooseYourPathContainer) {
-    const $slideTitle = a.querySelector('.choose-your-path-slide-title');
-    const slideName = $slideTitle ? textToName($slideTitle.textContent.trim()) : 'slide';
 
-    adobeEventName = `${adobeEventName}chooseYourPath:${slideName}:slidePressed`;
-  } else if ($contentToggleContainer) {
-    const toggleName = textToName(a.textContent.trim());
-    adobeEventName = `${adobeEventName}contentToggle:${toggleName}:buttonPressed`;
-  } else if (a.classList.contains('floating-button-lottie')) {
-    adobeEventName = `${adobeEventName}floatingButton:scrollPressed`;
-  } else if (a.classList.contains('video-player-inline-player-overlay')) {
-    const sessionName = a.parentNode.parentNode.parentNode.querySelector('.video-player-session-number').textContent.trim();
-    const videoName = a.parentNode.parentNode.parentNode.querySelector('.video-player-video-title').textContent.trim();
-    adobeEventName = `${adobeEventName}playing:${sessionName}-${videoName}`;
-  } else if (a.classList.contains('notch')) {
-    adobeEventName = `${adobeEventName}splitAction:notch`;
-  } else if (a.classList.contains('underlay')) {
-    adobeEventName = `${adobeEventName}splitAction:background`;
-  } else if (a.parentElement.classList.contains('floating-button')) {
-    adobeEventName = `${adobeEventName}floatingButton:ctaPressed`;
-  } else if (a.closest('.faq')) {
-    adobeEventName = appendLinkText(`${adobeEventName}faq:`, a);
-    // CTA in the hero
-  } else if (a.closest('.hero')) {
-    adobeEventName = appendLinkText(`${adobeEventName}hero:`, a);
-    // Click in the pricing block
-  } else if (sparkLandingPageType === 'express-your-fandom') {
-    adobeEventName = appendLinkText(`${adobeEventName}${sparkLandingPageType}:`, a);
-  } else if (sparkLandingPageType === 'express-your-brand') {
-    adobeEventName = appendLinkText(`${adobeEventName}learn:${sparkLandingPageType}:`, a);
-  } else if (sparkLandingPageType === 'pricing') {
-    if (a.tagName !== 'A') {
-      adobeEventName += `pricing:pricing:${a.textContent.trim()}:Click`;
-      // Creative cloud learn more
-    } else if (a.parentElement.id === 'adobe-spark-is-a-part-of-most-creative-cloud-paid-plans-learn-more') {
-      adobeEventName += 'pricing:creativeCloud:learnMore';
-    } else if (a.id === 'free-trial') {
-      adobeEventName += 'pricing:cta:StartForFree';
-    } else if (a.id === '3-month-trial') {
-      adobeEventName += 'pricing:cta:StartYour3MonthTrial';
-      // View plans
-    } else {
-      adobeEventName = 'adobe.com:express:CTA:pricing:viewPlans:Click';
+    // clicks using [data-lh and data-ll]
+    let trackingHeader = a.closest('[data-lh]');
+    if (trackingHeader || a.dataset.lh) {
+      adobeEventName = 'adobe.com:express';
+      let headerString = '';
+      while (trackingHeader) {
+        headerString = `:${textToName(trackingHeader.dataset.lh.trim())}${headerString}`;
+        trackingHeader = trackingHeader.parentNode.closest('[data-lh]');
+      }
+      adobeEventName += headerString;
+      if (a.dataset.ll) {
+        adobeEventName += `:${textToName(a.dataset.ll.trim())}`;
+      } else {
+        adobeEventName += `:${textToName(a.innerText.trim())}`;
+      }
     }
-    // quick actions clicks
-  } else if (a.closest('ccl-quick-action') && a.classList.contains('upload-your-photo')) {
-    // this event is handled at mock-file-input level
-    return;
-  } else if (a.href && (a.href.match(/spark\.adobe\.com\/[a-zA-Z-]*\/?tools/g) || a.href.match(/express\.adobe\.com\/[a-zA-Z-]*\/?tools/g))) {
-    adobeEventName = appendLinkText(adobeEventName, a);
-  } else if (a.href && (a.href.match(/spark\.adobe\.com\/[a-zA-Z-]*\/?tools/g) || a.href.match(/express\.adobe\.com\/[a-zA-Z-]*\/?express-apps\/animate-from-audio/g))) {
-    adobeEventName = appendLinkText(adobeEventName, a);
-    // Frictionless Quick Actions clicks
-  } else if (a.closest('ccl-quick-action') && (a.getAttribute('data-action') === 'Download')) {
-    adobeEventName = 'quickAction:downloadPressed';
-  } else if (a.closest('ccl-quick-action') && (a.getAttribute('data-action') === 'Editor')) {
-    adobeEventName = 'quickAction:openInEditorPressed';
-    // ToC clicks
-  } else if (a.closest('.toc-container')) {
-    if (a.classList.contains('toc-toggle')) {
-      adobeEventName += 'toc:toggle:Click';
-    } else if (a.classList.contains('toc-close')) {
-      adobeEventName += 'toc:close:Click';
-    } else if (a.classList.contains('toc-handle')) {
-      adobeEventName += 'toc:close:Click:handle';
-    } else if (a.classList.contains('toc-wrapper')) {
-      adobeEventName += 'toc:close:Click:background';
-    } else {
-      adobeEventName = appendLinkText(`${adobeEventName}toc:link:Click:`, a);
+    if (window.hlx?.experiment) {
+      let prefix = '';
+      if (window.hlx.experiment?.id) prefix = `${window.hlx.experiment.id}:`;
+      if (window.hlx.experiment?.selectedVariant) {
+        let variant = window.hlx.experiment.selectedVariant;
+        if (variant.includes('-')) [, variant] = variant.split('-');
+        prefix += `${variant}:`;
+      }
+      adobeEventName = prefix + adobeEventName;
     }
-  } else if (a.closest('.template')) {
-    adobeEventName = appendLinkText(adobeEventName, a);
-  } else if (a.closest('.tabs-ax .tab-list-container')) {
-    adobeEventName += `${a.closest('.tabs-ax')?.id}:${a.id}`;
-    // Default clicks
-  } else {
-    adobeEventName = appendLinkText(adobeEventName, a);
-  }
 
-  // clicks using [data-lh and data-ll]
-  let trackingHeader = a.closest('[data-lh]');
-  if (trackingHeader || a.dataset.lh) {
-    adobeEventName = 'adobe.com:express';
-    let headerString = '';
-    while (trackingHeader) {
-      headerString = `:${textToName(trackingHeader.dataset.lh.trim())}${headerString}`;
-      trackingHeader = trackingHeader.parentNode.closest('[data-lh]');
-    }
-    adobeEventName += headerString;
-    if (a.dataset.ll) {
-      adobeEventName += `:${textToName(a.dataset.ll.trim())}`;
-    } else {
-      adobeEventName += `:${textToName(a.innerText.trim())}`;
-    }
-  }
-  if (window.hlx?.experiment) {
-    let prefix = '';
-    if (window.hlx.experiment?.id) prefix = `${window.hlx.experiment.id}:`;
-    if (window.hlx.experiment?.selectedVariant) {
-      let variant = window.hlx.experiment.selectedVariant;
-      if (variant.includes('-')) [, variant] = variant.split('-');
-      prefix += `${variant}:`;
-    }
-    adobeEventName = prefix + adobeEventName;
-  }
-
-  _satellite.track('event', {
-    xdm: {},
-    data: {
-      eventType: 'web.webinteraction.linkClicks',
-      web: {
-        webInteraction: {
-          name: adobeEventName,
-          linkClicks: {
-            value: 1,
-          },
-          type: 'other',
-        },
-      },
-      _adobe_corpnew: {
-        digitalData: {
-          primaryEvent: {
-            eventInfo: {
-              eventName: adobeEventName,
+    _satellite.track('event', {
+      xdm: {},
+      data: {
+        eventType: 'web.webinteraction.linkClicks',
+        web: {
+          webInteraction: {
+            name: adobeEventName,
+            linkClicks: {
+              value: 1,
             },
+            type: 'other',
           },
-          ...(hemingwayAsset
-            ? {
-              asset: {
-                assetInfo: {
-                  assetId: hemingwayAssetId,
-                  assetPath: hemingwayAssetPath,
-                  assetPosition: hemingwayAssetPosition,
-                },
+        },
+        _adobe_corpnew: {
+          digitalData: {
+            primaryEvent: {
+              eventInfo: {
+                eventName: adobeEventName,
               },
-            }
-            : {}),
+            },
+            ...(hemingwayAsset
+              ? {
+                asset: {
+                  assetInfo: {
+                    assetId: hemingwayAssetId,
+                    assetPath: hemingwayAssetPath,
+                    assetPosition: hemingwayAssetPosition,
+                  },
+                },
+              }
+              : {}),
+          },
         },
       },
-    },
-  });
+    });
+  };
+  safelyFireAnalyticsEvent(fireEvent);
 }
 
 function trackVideoAnalytics(parameters) {
@@ -558,7 +540,7 @@ function decorateAnalyticsEvents() {
 
   // Tracking any link or links that is added after page loaded.
   d.addEventListener('linkspopulated', async (e) => {
-    await trackBranchParameters(e.detail);
+    trackBranchParameters(e.detail);
     e.detail.forEach(($link) => {
       $link.addEventListener('click', () => {
         trackButtonClick($link);
@@ -567,7 +549,7 @@ function decorateAnalyticsEvents() {
   });
 
   d.addEventListener('pricingdropdown', () => {
-    sendEventToAdobeAnaltics('adobe.com:express:pricing:bundleType:selected');
+    sendEventToAnalytics('adobe.com:express:pricing:bundleType:selected');
   });
 
   // tracking videos loaded asynchronously.
@@ -577,143 +559,33 @@ function decorateAnalyticsEvents() {
   });
 
   d.addEventListener('videoclosed', (e) => {
-    sendEventToAdobeAnaltics(`adobe.com:express:cta:learn:columns:${e.detail.parameters.videoId}:videoClosed`);
+    sendEventToAnalytics(`adobe.com:express:cta:learn:columns:${e.detail.parameters.videoId}:videoClosed`);
   });
 
   d.addEventListener('click', (e) => {
     if (e.target.id === 'mock-file-input') {
-      sendEventToAdobeAnaltics('adobe.com:express:cta:uploadYourPhoto');
+      sendEventToAnalytics('adobe.com:express:cta:uploadYourPhoto');
     }
   });
 }
 
 function martechLoadedCB() {
-  /* eslint-disable no-underscore-dangle */
-  //------------------------------------------------------------------------------------
-  // gathering the data
-  //------------------------------------------------------------------------------------
-
-  const locale = getConfig().locale.prefix;
-  const pathSegments = pathname.substr(1).split('/');
-  if (locale !== '') pathSegments.shift();
-  const pageName = `adobe.com:${pathSegments.join(':')}`;
-
-  const language = document.documentElement.getAttribute('lang');
-
-  let category = getMetadata('category');
-  if (!category && (pathname.includes('/create/')
-      || pathname.includes('/feature/'))) {
-    category = 'design';
-    if (pathname.includes('/image')) category = 'photo';
-    if (pathname.includes('/video')) category = 'video';
-  }
-
-  // home
-  if (
-    pathname === '/express'
-      || pathname === '/express/'
-  ) {
-    sparkLandingPageType = 'home';
-    // seo
-  } else if (
-    pathname === '/express/create'
-      || pathname.includes('/create/')
-      || pathname === '/express/make'
-      || pathname.includes('/make/')
-      || pathname === '/express/feature'
-      || pathname.includes('/feature/')
-      || pathname === '/express/discover'
-      || pathname.includes('/discover/')
-  ) {
-    sparkLandingPageType = 'seo';
-    // learn
-  } else if (
-    pathname === '/express/tools'
-      || pathname.includes('/tools/')
-  ) {
-    sparkLandingPageType = 'quickAction';
-  } else if (
-    pathname === '/express/learn'
-      || (
-        pathname.includes('/learn/')
-        && !pathname.includes('/blog/')
-      )
-  ) {
-    if (pathname.includes('/express-your-brand')) {
-      sparkLandingPageType = 'express-your-brand';
-    } else {
-      sparkLandingPageType = 'learn';
-    }
-    // blog
-  } else if (
-    pathname === '/express/learn/blog'
-      || pathname.includes('/learn/blog/')
-  ) {
-    sparkLandingPageType = 'blog';
-    // pricing
-  } else if (
-    pathname.includes('/pricing')
-  ) {
-    sparkLandingPageType = 'pricing';
-    // edu
-  } else if (
-    pathname.includes('/education/')
-  ) {
-    sparkLandingPageType = 'edu';
-    // other
-  } else if (
-    pathname.includes('/express-your-fandom')
-  ) {
-    sparkLandingPageType = 'express-your-fandom';
-  } else {
-    sparkLandingPageType = 'other';
-  }
-  const url = new URL(loc.href);
-  const sparkTouchpoint = url.searchParams.get('touchpointName');
-
-  //----------------------------------------------------------------------------
-  // set some global and persistent data layer properties
-  //----------------------------------------------------------------------------
-  set('page.pageInfo.pageName', pageName);
-  set('page.pageInfo.language', language);
-  set('page.pageInfo.siteSection', 'adobe.com:express');
-  set('page.pageInfo.category', category);
-
-  //----------------------------------------------------------------------------
-  // spark specific global and persistent data layer properties
-  //----------------------------------------------------------------------------
-  set('page.pageInfo.pageurl', loc.href);
-  set('page.pageInfo.namespace', 'express');
-
-  /* set experiment and variant information */
-  if (window.hlx.experiment) {
-    const { experiment } = window.hlx;
-    set('adobe.experienceCloud.target.info.primarytest.testinfo.campaignid', experiment.id);
-    set('adobe.experienceCloud.target.info.primarytest.testinfo.offerid', experiment.selectedVariant);
-  }
-
   //------------------------------------------------------------------------------------
   // Fire extra spark events
   //------------------------------------------------------------------------------------
 
   // Fire the viewedPage event
-  sendEventToAdobeAnaltics('viewedPage');
+  sendEventToAnalytics('viewedPage');
 
   // Fire the landing:viewedPage event
-  sendEventToAdobeAnaltics('landing:viewedPage');
-
-  // Fire quick-action-viewed event if needed
-  const quickActionBlock = d.querySelector('.frictionless-quick-action.block');
-  if (quickActionBlock) {
-    sendFrictionlessEventToAdobeAnaltics(quickActionBlock);
-  }
+  sendEventToAnalytics('landing:viewedPage');
 
   // Fire the displayPurchasePanel event if it is the pricing site
   if (
     sparkLandingPageType === 'pricing'
-      && sparkTouchpoint
+    && sparkTouchpoint
   ) {
-    sendEventToAdobeAnaltics('displayPurchasePanel');
+    sendEventToAnalytics('displayPurchasePanel');
   }
 
   const processed = {};
@@ -801,7 +673,7 @@ function martechLoadedCB() {
           audiences.push(ENABLE_PRICING_MODAL_AUDIENCE);
           segments.push(RETURNING_VISITOR_SEGMENT_ID);
 
-          sendEventToAdobeAnaltics('pricingModalUserInSegment');
+          sendEventToAnalytics('pricingModalUserInSegment');
         }
 
         QUICK_ACTION_SEGMENTS.forEach((QUICK_ACTION_SEGMENT) => {
@@ -828,6 +700,7 @@ function martechLoadedCB() {
 }
 
 export default async function initMartech() {
+  setDataAnalyticsAttributesForMartech();
   await loadScript(martechURL);
   return martechLoadedCB();
 }
